+++ packages/databricks-vscode/src/cli/CliWrapper.ts
@@ -41,10 +41,27 @@ function getValidHost(host: string) {
  * of the databricks CLI
  */
 export class CliWrapper {
-    constructor(private extensionContext: ExtensionContext) {}
+    constructor(
+        private extensionContext: ExtensionContext,
+        private logFilePath?: string
+    ) {}
 
     get cliPath(): string {
         return this.extensionContext.asAbsolutePath("./bin/databricks");
+    }
+
+    get loggingArguments(): string[] {
+        if (!workspaceConfigs.loggingEnabled) {
+            return [];
+        }
+        return [
+            "--log-level",
+            "debug",
+            "--log-file",
+            this.logFilePath ?? "stderr",
+            "--log-format",
+            "json",
+        ];
     }
 
     /**
@@ -61,12 +78,10 @@ export class CliWrapper {
             "--watch",
             "--output",
             "json",
+            ...this.loggingArguments,
         ];
         if (syncType === "full") {
             args.push("--full");
-        }
-        if (workspaceConfigs.cliVerboseMode) {
-            args.push("--log-level", "debug", "--log-file", "stderr");
         }
         return {command: this.cliPath, args};
     }
@@ -74,7 +89,12 @@ export class CliWrapper {
     private getListProfilesCommand(): Command {
         return {
             command: this.cliPath,
-            args: ["auth", "profiles", "--skip-validate"],
+            args: [
+                "auth",
+                "profiles",
+                "--skip-validate",
+                ...this.loggingArguments,
+            ],
         };
     }
 
@@ -83,17 +103,35 @@ export class CliWrapper {
         configfilePath?: string,
         @context ctx?: Context
     ): Promise<Array<ConfigEntry>> {
-        const cmd = await this.getListProfilesCommand();
-        const res = await execFile(cmd.command, cmd.args, {
-            env: {
-                /*  eslint-disable @typescript-eslint/naming-convention */
-                HOME: process.env.HOME,
-                DATABRICKS_CONFIG_FILE:
-                    configfilePath || process.env.DATABRICKS_CONFIG_FILE,
-                DATABRICKS_OUTPUT_FORMAT: "json",
-                /*  eslint-enable @typescript-eslint/naming-convention */
-            },
-        });
+        const cmd = this.getListProfilesCommand();
+
+        let res;
+        try {
+            res = await execFile(cmd.command, cmd.args, {
+                env: {
+                    /*  eslint-disable @typescript-eslint/naming-convention */
+                    HOME: process.env.HOME,
+                    DATABRICKS_CONFIG_FILE:
+                        configfilePath || process.env.DATABRICKS_CONFIG_FILE,
+                    DATABRICKS_OUTPUT_FORMAT: "json",
+                    /*  eslint-enable @typescript-eslint/naming-convention */
+                },
+            });
+        } catch (e) {
+            let msg = "Failed to load Databricks Config File";
+            if (e instanceof Error) {
+                if (e.message.includes("cannot parse config file")) {
+                    msg =
+                        "Failed to parse Databricks Config File, please make sure it's in the correct ini format";
+                } else if (e.message.includes("spawn UNKNOWN")) {
+                    msg = `Failed to parse Databricks Config File using databricks CLI, please make sure you have permissions to execute this binary: "${this.cliPath}"`;
+                }
+            }
+            ctx?.logger?.error(msg, e);
+            this.showConfigFileWarning(msg);
+            return [];
+        }
+
         const profiles = JSON.parse(res.stdout).profiles || [];
         const result = [];
 
@@ -115,22 +153,24 @@ export class CliWrapper {
                     msg = `Error parsing profile ${profile.name}`;
                 }
                 ctx?.logger?.error(msg, e);
-                window
-                    .showWarningMessage(
-                        msg,
-                        "Open Databricks Config File",
-                        "Ignore"
-                    )
-                    .then((choice) => {
-                        if (choice === "Open Databricks Config File") {
-                            commands.executeCommand(
-                                "databricks.connection.openDatabricksConfigFile"
-                            );
-                        }
-                    });
+                this.showConfigFileWarning(msg);
             }
         }
         return result;
+    }
+
+    private async showConfigFileWarning(msg: string) {
+        const openAction = "Open Databricks Config File";
+        const choice = await window.showWarningMessage(
+            msg,
+            openAction,
+            "Ignore"
+        );
+        if (choice === openAction) {
+            commands.executeCommand(
+                "databricks.connection.openDatabricksConfigFile"
+            );
+        }
     }
 
     public async getBundleSchema(): Promise<string> {