+++ packages/databricks-vscode/src/configuration/configureWorkspaceWizard.ts
@@ -72,6 +72,7 @@ export async function configureWorkspaceWizard(
                 return false;
             },
             items,
+            ignoreFocusOut: true,
         });
 
         state.host = normalizeHost(host);
@@ -155,6 +156,7 @@ export async function configureWorkspaceWizard(
             totalSteps: 2,
             placeholder: "Select authentication method",
             items,
+            ignoreFocusOut: true,
             shouldResume: async () => {
                 return false;
             },